feat: full-text search over captured command output
Add the SearchCommandOutput gRPC RPC (streaming OutputSearchMatch results
with highlighted snippets), the daemon client method, and the
`atuin search-output` CLI subcommand. The SearchIndex limit type is
normalized to usize.

+++ crates/atuin-search-bench/benches/search.rs
@@ -48,7 +48,7 @@ const QUERIES: &[&str] =
     &["", "g", "git", "git p", "cargo build", "docker compose up", "zzznomatchzzz"];
 
 /// The interactive UI requests up to 200 results per query.
-const LIMIT: u32 = 200;
+const LIMIT: usize = 200;
 
 /// Working directories assigned round-robin to history entries, so the
 /// directory-filtered benchmark has a realistic candidate subset.
